backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS
psycopg2-binary
requests
python-dotenv
marshmallow
google-generativeai
